Add Hyper-V support
Currently only tested on the host end but it works correctly when
communicating with the client side test harness under WSL.
