+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "vsock"
 version = "0.2.6"
-authors = ["fsyncd", "rust-vsock"]
+authors = ["fsyncd", "rust-vsock", "reanimus"]
 description = "Virtio socket support for Rust"
 repository = "https://github.com/rust-vsock/vsock-rs"
 homepage = "https://github.com/rust-vsock/vsock-rs"
@@ -12,8 +12,36 @@ exclude = ["test_fixture"]
 
 [dependencies]
 libc = "0.2.79"
+
+[target.'cfg(unix)'.dependencies]
 nix = "0.23.0"
+
+[target.'cfg(windows)'.dependencies]
+widestring = "0.5.1"
+
+[target.'cfg(windows)'.dependencies.windows]
+version = "0.35.0"
+features = [
+    "Win32_Foundation",
+    "Win32_Networking_WinSock",
+    "Win32_Security",
+    "Win32_System_Diagnostics_Debug",
+    "Win32_System_Hypervisor",
+    "Win32_System_Memory",
+]
 
 [dev-dependencies]
 rand = "0.8.3"
 sha2 = ">=0.8.0"
+
+
+[target.'cfg(windows)'.dev-dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1"
+uuid = { version = "0.8", features = ["serde"] }
+
+[target.'cfg(windows)'.dev-dependencies.windows]
+version = "0.35.0"
+features = [
+    "Win32_System_HostComputeSystem",
+]
