+++ tests/vsock.rs
@@ -15,9 +15,59 @@
  */
 
 use rand::RngCore;
+use serde::Deserialize;
 use sha2::{Digest, Sha256};
 use std::io::{Read, Write};
+use std::ptr;
+
+#[cfg(target_os = "unix")]
 use vsock::{get_local_cid, SockAddr, VsockAddr, VsockStream, VMADDR_CID_HOST};
+
+#[cfg(target_os = "windows")]
+use vsock::{svcid_from_port, VsockStream};
+#[cfg(target_os = "windows")]
+use widestring::U16CString;
+#[cfg(target_os = "windows")]
+use windows::core::GUID;
+#[cfg(target_os = "windows")]
+use windows::core::PCWSTR;
+#[cfg(target_os = "windows")]
+use windows::Win32::System::HostComputeSystem::{
+    HcsCreateOperation, HcsEnumerateComputeSystems, HcsWaitForOperationResult,
+};
+#[cfg(target_os = "windows")]
+use windows::Win32::System::Memory::LocalFree;
+#[cfg(target_os = "windows")]
+use uuid::Uuid;
+
+#[cfg(target_os = "windows")]
+#[allow(dead_code)]
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+struct ComputeSystem {
+    pub id: Uuid,
+    pub system_type: String,
+    pub owner: String,
+    pub runtime_id: Uuid,
+    pub state: String,
+}
+
+#[cfg(target_os = "windows")]
+unsafe fn get_first_vm() -> GUID {
+    let op = HcsCreateOperation(ptr::null(), None);
+    if op.is_invalid() {
+        panic!("Failed to create operation!");
+    }
+    HcsEnumerateComputeSystems(PCWSTR(ptr::null()), op).expect("Failed to enumerate systems");
+    let result =
+        HcsWaitForOperationResult(op, u32::MAX).expect("Failed to wait for operation to complete!");
+    let json_str = U16CString::from_ptr_str(result.0).to_string_lossy();
+    LocalFree(result.0 as isize);
+    let systems: Vec<ComputeSystem> =
+        serde_json::from_str(&json_str).expect("Failed to decode HCS JSON");
+    let machine = systems.first().expect("No machines found");
+    GUID::from_u128(machine.id.as_u128())
+}
 
 const TEST_BLOB_SIZE: usize = 1_000_000;
 const TEST_BLOCK_SIZE: usize = 5_000;
@@ -39,8 +89,13 @@ fn test_vsock() {
     rx_blob.resize(TEST_BLOB_SIZE, 0);
     rng.fill_bytes(&mut blob);
 
+    #[cfg(target_os = "unix")]
     let mut stream =
         VsockStream::connect(&SockAddr::Vsock(VsockAddr::new(3, 8000))).expect("connection failed");
+    #[cfg(target_os = "windows")]
+    let mut stream =
+        VsockStream::connect_with_vmid_svcid(unsafe { get_first_vm() }, svcid_from_port(8000))
+            .expect("connection failed");
 
     while tx_pos < TEST_BLOB_SIZE {
         let written_bytes = stream
@@ -69,6 +124,7 @@ fn test_vsock() {
 }
 
 #[test]
+#[cfg(target_os = "unix")]
 fn test_get_local_cid() {
     assert_eq!(get_local_cid().unwrap(), VMADDR_CID_HOST);
 }
